rename activation_function and initialization_functions
